requirements.txt
torch
torchvision
opencv-python
numpy
pillow
matplotlib
tqdm